nit: ServiceConnectionContainerTests
